mindmy_ai_full/pyproject.toml
[tool.poetry]
name = "mindmy-ai"
version = "0.1.0"
description = "MindMy AI — cognitive operating system"
authors = ["MindMy Team <dev@example.com>"]
readme = "README.md"
packages = [{ include = "mindmy_*" }]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.30"}
python-jose = "^3.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
supabase-py = "^2.3"
redis = "^5.0"
alembic = "^1.13"
openai = "^1.30"
httpx = "^0.27"
loguru = "^0.7"
sentry-sdk = "^2.2"
prometheus-client = "^0.20"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
coverage = "^7.5"
ruff = "^0.4.6"
black = "^24.4"
mypy = "^1.10"
pre-commit = "^3.7"

[tool.poetry.scripts]
mindmy = "cli.main:_run"
